+++ utils.cc
@@ -3,20 +3,20 @@
 #include"object.hh"
 #include"utils.hh"
 
-double pointdirection(double x0,double y0,double x1,double y1){
-  double r=atan2(y1-y0,x1-x0);
+float pointdirection(float x0,float y0,float x1,float y1){
+  float r=atan2(y1-y0,x1-x0);
   return radtodeg(r<0?r+2*PIE:r);
 }
-double pointdistance(double x0,double y0,double x1,double y1){
+float pointdistance(float x0,float y0,float x1,float y1){
   return sqrt(pow(x1-x0,2)+pow(y1-y0,2));
 }
-double degtorad(double x){
+float degtorad(float x){
   return x/180*PIE;
 }
-double radtodeg(double x){
+float radtodeg(float x){
   return x*180/PIE;
 }
-int sign(double x){  
+int sign(float x){  
   return (x>0)-(x<0);
 }
 int arbitrary(int i){
