+++ utils.hh
@@ -4,9 +4,9 @@
 #include<random>
 #include"env.hh"
 
-double pointdirection(double,double,double,double);
-double pointdistance (double,double,double,double);
-double degtorad(double);
-double radtodeg(double);
-int    sign(double);
-int    arbitrary(int);
+float pointdirection(float,float,float,float);
+float pointdistance (float,float,float,float);
+float degtorad(float);
+float radtodeg(float);
+int   sign(float);
+int   arbitrary(int);
